feature: dialog-prompt + admin-page + quiz tests

See merge request polytechnique-montr-al/log2990/20241/equipe-206/LOG2990-206!38

+++ client/src/app/components/quiz/quiz.component.ts
@@ -45,9 +45,12 @@ export class QuizComponent {
 
     deleteQuiz() {
         // eslint-disable-next-line no-underscore-dangle
-        this.quizHttpService.deleteQuizById(this.quiz._id).subscribe(() => {
-            this.isDeleted = true;
-        });
+        if (this.quizHttpService.deleteQuizById(this.quiz._id)) {
+            // eslint-disable-next-line no-underscore-dangle
+            this.quizHttpService.deleteQuizById(this.quiz._id).subscribe(() => {
+                this.isDeleted = true;
+            });
+        }
     }
 
     editQuiz() {
